Chore Album CRUD by adding validation

+++ backend/src/services/songs.service.ts
@@ -1,12 +1,15 @@
 import { Repository } from "typeorm";
 import Song from "../entities/songs.entity";
 import dbConn from "../database/postgresConnection";
+import Album from "../entities/albuns.entity";
 
 class SongService {
     songRepository: Repository<Song>;
+    albumRepository: Repository<Album>;
 
     constructor() {
         this.songRepository = dbConn.getRepository(Song);
+        this.albumRepository = dbConn.getRepository(Album);
     }
 
     async getSongById(id: number): Promise<Song> {
@@ -23,7 +26,21 @@ class SongService {
         return await this.songRepository.find({ relations: ["album"] });
     }
 
-    async deleteSongsByAlbumId(albumId: number) {
+    async deleteSongsByAlbumId(albumId: number, artist_login: string) {
+        const album = await this.albumRepository.findOne({ 
+            where: { albumID: albumId }, 
+            relations: ["artist"] 
+        });
+    
+        if (!album) {
+            throw new Error("Album not found");
+        }
+    
+        if (album.artist.login !== artist_login) {
+            console.log(album.artist.login)
+            throw new Error("You do not have permission to delete songs from this album");
+        }
+    
         await this.songRepository.delete({ album: { albumID: albumId } }); 
     }
